Implementar modulo de Inventario y Compras

+++ src/main/java/pe/edu/utp/huellitas/repository/RolRepository.java
@@ -5,7 +5,7 @@ import pe.edu.utp.huellitas.model.Rol;
 
 import java.util.Optional;
 
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
 
     Optional<Rol> findByNombre(String nombre);
 }
